Add files via upload
Problems
Using the pop-up prompt
Prompt the user for a width for a rectangle
Prompt the user for a height for a rectangle
Alert the user to the resulting area and perimeter of the resulting rectangle. The alert should also include the user-provided width and height information.
Using a form
Start with the form noted below
Complete the simple function getarea, outlined below, to compute the area and perimeter of the rectangle
Display the result to the user by inserting the area result in the element with the output with the area id and the perimeter result in the output with the perim id.